refactor: rename store to repository pkg
