+++ domain/command.go
@@ -16,7 +16,7 @@ type CommandService interface {
 }
 
 // Command repository
-type CommandStore interface {
+type CommandRepository interface {
 	Add(ctx context.Context, command *Commmand) error
 	Get(ctx context.Context, key string) (*Commmand, error)
 	List(ctx context.Context, query string) ([]string, error)
